greedy revisado, falta probarlo

+++ algoritmos.h
@@ -1,7 +1,7 @@
 #ifndef PRACTICA1_MH_PAR_ALGORITMOS_H
 #define PRACTICA1_MH_PAR_ALGORITMOS_H
 #include <random>
-#include <limits>
+#include <cfloat>
 #include "utilidades.h"
 
 // cj es igual al cluster al que pertenece xk y la restricción es cannot-link
@@ -17,10 +17,12 @@ size_t infeasibility(size_t xi, int cj, const std::vector<int> C, const R_matrix
     return inf;
 }
 
-std::vector<size_t> greedy_copkm(const double_matrix& X, const R_matrix& R, std::vector<cluster>& clusters, size_t seed)
+std::vector<int> greedy_copkm(const double_matrix& X, const R_matrix& R, std::vector<cluster>& clusters, size_t seed)
 {
     std::vector<int> C(X.size(), -1); // -1 == ningún cluster, 0 == 1º cluster, 1 == 2º cluster....
     std::vector<size_t> vRSI;
+
+    // Bajaramos los índices para recorrer X de forma aleatoria y sin repetición
     for (size_t i = 0; i < X.size(); ++i)
         vRSI.push_back(i);
     shuffle(vRSI.begin(), vRSI.end(), std::default_random_engine(seed));
@@ -38,6 +40,7 @@ std::vector<size_t> greedy_copkm(const double_matrix& X, const R_matrix& R, std:
             min = 65535;
             asignaciones.clear();
 
+            // Calcular el incremento en infeasibility que produce la asignacion xi a cada cluster cj
             for (size_t cj = 0; cj < clusters.size(); ++cj) {
                 inf = infeasibility(xi, cj, C, R);
                 if (inf < min) {
@@ -48,6 +51,8 @@ std::vector<size_t> greedy_copkm(const double_matrix& X, const R_matrix& R, std:
                     asignaciones.push_back(cj);
             }
 
+            // De entre las asignaciones que produce nmenos incremento en infeasibility, seleccionar la
+            // asociada al centroide más cercano a xi
             for (auto& cj : asignaciones) {
                 d = clusters[cj].distancia_centroide(X[xi]);
                 if (d < d_min) {
@@ -60,6 +65,7 @@ std::vector<size_t> greedy_copkm(const double_matrix& X, const R_matrix& R, std:
             C[xi] = c_min;
         }
 
+        // Actualizar el centroide promediando las instancias de su cluster asociado ci
         for (auto& ci : clusters)
             ci.actualizar_centroide();
